refactor(analysis-types): extract baseline complexity section DTO
